created the tracker class and moved all the methods in tracker
